requirements.txt
Flask
connexion[swagger-ui]
matplotlib
html-table-parser-python3
pandas
urllib3
tables
pickle5